fix: update.

+++ lab1/include/Constants.h
@@ -5,7 +5,7 @@
 
 namespace Constants {
 	const int M = 8;
-	const int sizes[M] = {1000, 2000, 4000, 8000, 16000, 32000, 64000, 128000};
+	const long long sizes[M] = {1000, 2000, 4000, 8000, 16000, 32000, 64000, 128000};
 
 	const int amount_of_attempts = 20;
 
@@ -14,7 +14,7 @@ namespace Constants {
 
 	const std::string folder("plots/files/");
 
-	const double c = 0.6 * 1e-5;
+	const long double c = 0.6 * 1e-5;
 }
 
 #endif
